requirements.txt
bidict==0.22.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
pymongo==4.5.0
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
requests==2.31.0
simple-websocket==1.0.0
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
